requirements.txt
streamlit

langchain
langchain-core
langchain-community
langchain-text-splitters

langchain-google-genai

faiss-cpu

pypdf
python-dotenv
